Reusing and simplifying code
