+++ src/goFormat.ts
@@ -10,7 +10,7 @@ import cp = require('child_process');
 import path = require('path');
 import dmp = require('diff-match-patch');
 import { getBinPath } from './goPath';
-import { EditTypes, Edit } from './util';
+import { EditTypes, Edit, GetEditsFromDiffs } from './util';
 
 export class Formatter {
 	private formatCommand = 'goreturns';
@@ -39,62 +39,18 @@ export class Formatter {
 					let d = new dmp.diff_match_patch();
 
 					let diffs = d.diff_main(document.getText(), text);
-					let line = 0;
-					let character = 0;
-					let edits: vscode.TextEdit[] = [];
-					let edit: Edit = null;
+					let edits: Edit[] = GetEditsFromDiffs(diffs, 0);
+					let textEdits: vscode.TextEdit[] = [];
+
+					if (!edits){
+						return reject('Cannot format due to internal errors');
+					}
 
 					for (let i = 0; i < diffs.length; i++) {
-						let start = new vscode.Position(line, character);
-
-						// Compute the line/character after the diff is applied.
-						for (let curr = 0; curr < diffs[i][1].length; curr++) {
-							if (diffs[i][1][curr] !== '\n') {
-								character++;
-							} else {
-								character = 0;
-								line++;
-							}
-						}
-
-						switch (diffs[i][0]) {
-							case dmp.DIFF_DELETE:
-								if (edit == null) {
-									edit = new Edit(EditTypes.EDIT_DELETE, start);
-								} else if (edit.action !== EditTypes.EDIT_DELETE) {
-									return reject('cannot format due to an internal error.');
-								}
-								edit.end = new vscode.Position(line, character);
-								break;
-
-							case dmp.DIFF_INSERT:
-								if (edit == null) {
-									edit = new Edit(EditTypes.EDIT_INSERT, start);
-								} else if (edit.action === EditTypes.EDIT_DELETE) {
-									edit.action = EditTypes.EDIT_REPLACE;
-								}
-								// insert and replace edits are all relative to the original state
-								// of the document, so inserts should reset the current line/character
-								// position to the start.		
-								line = start.line;
-								character = start.character;
-								edit.text += diffs[i][1];
-								break;
-
-							case dmp.DIFF_EQUAL:
-								if (edit != null) {
-									edits.push(edit.apply());
-									edit = null;
-								}
-								break;
-						}
+						textEdits.push(edits[i].apply());
 					}
 
-					if (edit != null) {
-						edits.push(edit.apply());
-					}
-
-					return resolve(edits);
+					return resolve(textEdits);
 				} catch (e) {
 					reject(e);
 				}
